refactor(CreateMySQLDBConn): move CreateMySQLDBConn function from db_helper.go to factory.go
